moved test code into test_sum_divide.py and included checks for ZeroDivisionError
